pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["old*"]

[tool.setuptools.package-data]
byzantine = ["data/*"]

[project]
name = "byzantine"
version = "0.1.0"
description = "A Python library"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = ["python-dateutil", "pyyaml", "docopt"]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
fasting-calendar = "byzantine.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
